LayoutTests/js/dom/string-fontsize-expected.txt

This is a test case for String.prototype.fontsize(size).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS '_'.fontsize('"') is "<font size=\"&quot;\">_</font>"
PASS '_'.fontsize('b') is "<font size=\"b\">_</font>"
PASS '<'.fontsize('b') is "<font size=\"b\"><</font>"
PASS '_'.fontsize(0x2A) is "<font size=\"42\">_</font>"
PASS '_'.fontsize('"') is "<font size=\"&quot;\">_</font>"
PASS '_'.fontsize('" color="b') is "<font size=\"&quot; color=&quot;b\">_</font>"
PASS String.prototype.fontsize.call(0x2A, 0x2A) is "<font size=\"42\">42</font>"
PASS String.prototype.fontsize.call(undefined) threw exception TypeError: Type error.
PASS String.prototype.fontsize.call(null) threw exception TypeError: Type error.
PASS String.prototype.fontsize.length is 1
PASS successfullyParsed is true

TEST COMPLETE
